modified parsing unit tests
